backend/pyproject.toml
[project]
name = "blinky-backend"
version = "1.0.0"
description = "Blinky - Minimal Bookmark Manager API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "sqlmodel>=0.0.31",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "authlib>=1.3.2",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.5.0",
]
